+++ apps/editing-toolkit/editing-toolkit-plugin/full-site-editing-plugin.php
@@ -321,7 +321,7 @@ add_action( 'plugins_loaded', __NAMESPACE__ . '\load_wpcom_block_editor_sidebar'
  */
 function load_coming_soon() {
 	if ( defined( 'WPCOM_PUBLIC_COMING_SOON' ) && WPCOM_PUBLIC_COMING_SOON ) {
-		require_once __DIR__ . '/wpcom-coming-soon/class-wpcom-coming-soon.php';
+		require_once __DIR__ . '/coming-soon/coming-soon.php';
 	}
 }
 add_action( 'plugins_loaded', __NAMESPACE__ . '\load_coming_soon' );
