--- apps/editing-toolkit/editing-toolkit-plugin/wpcom-coming-soon/README.md
@@ -1,4 +1,0 @@
-# Coming Soon
-
-Description coming soon...
-
