+++ apps/editing-toolkit/package.json
@@ -66,9 +66,9 @@
 		"wpcom-block-editor-nav-sidebar": "calypso-build --source='wpcom-block-editor-nav-sidebar'",
 		"dev:wpcom-block-editor-nav-sidebar": "yarn run wpcom-block-editor-nav-sidebar",
 		"build:wpcom-block-editor-nav-sidebar": "NODE_ENV=production yarn run wpcom-block-editor-nav-sidebar",
-		"wpcom-coming-soon": "calypso-build --source='wpcom-coming-soon'",
-		"dev:wpcom-coming-soon": "yarn run wpcom-coming-soon",
-		"build:wpcom-coming-soon": "NODE_ENV=production yarn run wpcom-coming-soon",
+		"coming-soon": "calypso-build --source='coming-soon'",
+		"dev:coming-soon": "yarn run coming-soon",
+		"build:coming-soon": "NODE_ENV=production yarn run coming-soon",
 		"dev": "node bin/npm-run-build.js --dev",
 		"build": "node bin/npm-run-build.js --build",
 		"postbuild": "yarn validate-es5",
